gofr works, corrected integer overflow bug
git-svn-id: svn://innoue/tracking/trunk@80 4cb6fdf0-3dc9-4801-8ffc-c6659000aa44

+++ cxx/corr_gofr.cpp
@@ -142,8 +142,9 @@ void Corr_gofr::display()const
 
 void Corr_gofr::normalize(vector<float> & out)const
 {
-
-  int count_sum = 0;
+  // tac 2009-10-09
+  // changed to float to prevent 
+  float count_sum = 0;
   for(int j = 0; j<n_bins_;++j)
     count_sum += bin_count_[j];
